[ADD] FutureDetailsFragment + ViewModel, add to model and DB Rain/Snow fields and others + [ERROR] adapter is not attached to layout in FutureListWeatherFragment

+++ app/src/main/java/com/bzahov/weatherapp/data/db/entity/forecast/model/Main.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class Main(
     @SerializedName("grnd_level")
     val grndLevel: Double,
-    val humidity: Int,
+    val humidity: Double,
     @SerializedName("feels_like")
     val feelsLike : Double,
     val pressure: Double,
